Fixed a race condition when shutting down Agg
When the Agg is stopped, the `done` channel is written to, which signals
its `aggregate()` goroutine loop to exit. This is fine, except this `done`
channel is also returned to the caller via `Done()`. The intended flow
is for `aggregate()` to catch the `done` channel write, and then wrap
up, close the output channel that the caller is reading from, and write
again to the `done` channel for the caller. However, if the caller
catches the first message before Agg does, then Agg never shuts down.

The fix here is to have two channels - one to signal that we're shutting
down (`quitting`), and one to signal that we've shut down (`done`). The
first one is private, just for communication with its own goroutines.
The second is exposed to the caller. We're also now closing channels,
rather than writing to them, for good measure, since we're only using
them as quit channels.